Convert utils to typescript
