--- app/utils/index.js
@@ -1,27 +1,0 @@
-import findTechDetails from './findTechDetails.js';
-import checkOffer from './checkOffer.js';
-import requestCurrentOffer from './requestCurrentOffer.js';
-import getOfferPageHtml from './getOfferPageHtml.js';
-import getOfferHtmlVals from './getOfferHtmlVals.js';
-import calculateSavings from './calculateSavings.js';
-import getOffer from './getOffer.js';
-import checkForLatestOffer from './checkForLatestOffer.ts';
-import logTool from './logTool.js';
-import logOffer from './logOffer.js';
-import soundAlert from './soundAlert.js';
-import storeOffer from './storeOffer.js';
-
-export {
-    findTechDetails,
-    checkOffer,
-    requestCurrentOffer,
-    getOfferPageHtml,
-    getOfferHtmlVals,
-    calculateSavings,
-    getOffer,
-    checkForLatestOffer,
-    logTool,
-    logOffer,
-    soundAlert,
-    storeOffer
-};
