--- app/utils/requestCurrentOffer.js
@@ -1,8 +1,0 @@
-import { getOfferPageHtml } from './index.js';
-
-const requestCurrentOffer = async () => {
-    const offerPageRes = await getOfferPageHtml();
-    return await offerPageRes.text();
-}
-
-export default requestCurrentOffer;
