--- app/utils/soundAlert.js
@@ -1,7 +1,0 @@
-import { logTool } from "./index.js";
-
-const soundAlert = () => {
-    logTool({ msg: '\u0007' });
-};
-
-export default soundAlert;
