+++ app/utils/soundAlert.ts
@@ -1,0 +1,7 @@
+import { logTool } from ".";
+
+const soundAlert = (): void => {
+    logTool({ msg: ['\u0007'] });
+};
+
+export default soundAlert;
