+++ app/utils/storeOffer.ts
@@ -1,0 +1,28 @@
+import { writeDB } from '../json_db/db.js';
+import { logTool } from '.';
+import { Offer } from '../../types';
+
+interface StoreOffer {
+    offer: Offer;
+}
+
+const storeOffer = async ({ offer }: StoreOffer): Promise<void> => {
+    try {
+        await writeDB({
+            name: offer.name,
+            price: offer.price,
+            web_price: offer.webPrice, 
+            retail_price: offer.retailPrice,
+            description: offer.descriptionText, 
+            points_ratings: offer.pointsRatings,
+            tech_details: offer.techDetails,
+            created: new Date(),
+            updated: new Date()
+        });
+    } catch (err) {
+        logTool({ color: 'red', msg: ['Error storing result in db', err] });
+        throw err;
+    }
+};
+
+export default storeOffer;
